Test that we did not commit to. Will change when project is more complete

+++ app/src/main/java/com/example/project02_triviaapp/database/CategoryDAO.java
@@ -29,6 +29,7 @@ public interface CategoryDAO {
     LiveData<Category> getCategoryByCategoryText(String category);
 
     @Query("SELECT * FROM " + TriviaDatabase.CATEGORY_TABLE + " WHERE categoryId = :categoryId")
-    LiveData<Category> getCategoryByCategoryId(int categoryId);
+    LiveData<Category> getCategoryByCategoryId(long categoryId);
+
 
 }
